+++ src/minimap/minimap.h
@@ -6,7 +6,7 @@
 /*   By: rhorbach <rhorbach@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/03/20 14:27:29 by rhorbach      #+#    #+#                 */
-/*   Updated: 2024/03/20 18:46:57 by rhorbach      ########   odam.nl         */
+/*   Updated: 2024/03/22 15:37:26 by rhorbach      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@
 
 # define MINIMAP_SCALE 20
 
-void	draw_minimap_ray(t_data *data, int end_x, int end_y);
+void	draw_minimap_ray(t_data *data, double end_x, double end_y);
 
 void	draw_minimap(t_data *data);
